Move semantic search related stuff to separate module
