+++ server/bleep/src/indexes/file.rs
@@ -8,15 +8,6 @@ use std::{
 use anyhow::{Context, Result};
 use async_trait::async_trait;
 use dashmap::mapref::entry::Entry;
-use maplit::hashmap;
-use ndarray_linalg::norm::Norm;
-use qdrant_client::{
-    prelude::{QdrantClient, QdrantClientConfig},
-    qdrant::{
-        point_id::PointIdOptions, vectors_config, CreateCollection, Distance, PointId, PointStruct,
-        VectorParams, VectorsConfig,
-    },
-};
 use tantivy::{
     collector::TopDocs,
     doc,
@@ -30,7 +21,6 @@ use tantivy::{
 use tokenizers as _;
 use tokio::runtime::Handle;
 use tracing::{debug, info, trace, warn};
-use tract_onnx::{prelude::*, tract_hir::internal::InferenceOp};
 
 #[cfg(feature = "debug")]
 use {
@@ -43,9 +33,9 @@ use super::{
     DocumentRead, Indexable, Indexer,
 };
 use crate::{
-    chunk,
     ctags::ctags_for_file,
     intelligence::TreeSitterFile,
+    semantic::Semantic,
     state::{FileCache, RepoHeadInfo, RepoRef, Repository},
     symbol::SymbolLocations,
     Configuration,
@@ -64,9 +54,7 @@ struct Workload<'a> {
 pub struct File {
     config: Arc<Configuration>,
     schema: Schema,
-    qdrant: Arc<QdrantClient>,
-    tokenizer: Arc<tokenizers::Tokenizer>,
-    onnx: Arc<InferenceSimplePlan<Graph<InferenceFact, Box<dyn InferenceOp + 'static>>>>,
+    semantic: Semantic,
 
     #[cfg(feature = "debug")]
     histogram: Arc<RwLock<Histogram>>,
@@ -139,23 +127,6 @@ impl File {
         let raw_content = builder.add_bytes_field("raw_content", FAST);
         let raw_repo_name = builder.add_bytes_field("raw_repo_name", FAST);
         let raw_relative_path = builder.add_bytes_field("raw_relative_path", FAST);
-        let qdrant = QdrantClient::new(Some(QdrantClientConfig::from_url(&config.qdrant_url)))
-            .await
-            .unwrap();
-
-        // don't panic if the collection already exists
-        _ = qdrant
-            .create_collection(&CreateCollection {
-                collection_name: "documents".to_string(),
-                vectors_config: Some(VectorsConfig {
-                    config: Some(vectors_config::Config::Params(VectorParams {
-                        size: 384,
-                        distance: Distance::Cosine.into(),
-                    })),
-                }),
-                ..Default::default()
-            })
-            .await;
 
         Self {
             file_disk_path,
@@ -171,19 +142,7 @@ impl File {
             avg_line_length,
             last_commit_unix_seconds,
             schema: builder.build(),
-            qdrant: qdrant.into(),
-            tokenizer: tokenizers::Tokenizer::from_file(Path::join(
-                &config.model_dir,
-                "tokenizer.json",
-            ))
-            .unwrap()
-            .into(),
-            onnx: onnx()
-                .model_for_path(Path::join(&config.model_dir, "model.onnx"))
-                .unwrap()
-                .into_runnable()
-                .unwrap()
-                .into(),
+            semantic: Semantic::new(&config.model_dir, &config.qdrant_url).await,
             config,
             raw_content,
             raw_repo_name,
@@ -514,112 +473,12 @@ impl File {
         let lines_avg = buffer.len() as f64 / buffer.lines().count() as f64;
         let last_commit = repo_info.last_commit_unix_secs;
 
-        let chunks = chunk::tree_sitter(&buffer, lang_str).unwrap_or_else(|e| {
-            debug!(?e, %lang_str, "failed to chunk, falling back to trivial chunker");
-            chunk::trivial(&buffer, 15) // line-wise chunking, 15 lines per chunk
+        _ = Handle::current().block_on(async {
+            self.semantic
+                .insert_points_for_buffer(&file_disk_path, &buffer, lang_str)
         });
-        debug!("found {} chunks", chunks.len());
-        let datapoints = chunks
-            .par_iter()
-            .filter_map(|&chunk| {
-                trace!("ndarray");
-                debug!(
-                    "chunk size: {}b {}",
-                    chunk.len(),
-                    (chunk.len() > 800)
-                        .then_some("(big chunk)")
-                        .unwrap_or_default()
-                );
-                let tokenizer_output = self.tokenizer.encode(chunk, true).unwrap();
-
-                let input_ids = tokenizer_output.get_ids();
-                let attention_mask = tokenizer_output.get_attention_mask();
-                let token_type_ids = tokenizer_output.get_type_ids();
-
-                let length = input_ids.len();
-
-                let input_ids: Tensor = tract_ndarray::Array2::from_shape_vec(
-                    (1, length),
-                    input_ids.iter().map(|&x| x as i64).collect(),
-                )
-                .map_err(|e| {
-                    tracing::error!("{e}");
-                    e
-                })
-                .ok()?
-                .into();
-                let attention_mask: Tensor = tract_ndarray::Array2::from_shape_vec(
-                    (1, length),
-                    attention_mask.iter().map(|&x| x as i64).collect(),
-                )
-                .map_err(|e| {
-                    tracing::error!("{e}");
-                    e
-                })
-                .ok()?
-                .into();
-                let token_type_ids: Tensor = tract_ndarray::Array2::from_shape_vec(
-                    (1, length),
-                    token_type_ids.iter().map(|&x| x as i64).collect(),
-                )
-                .map_err(|e| {
-                    tracing::error!("{e}");
-                    e
-                })
-                .ok()?
-                .into();
-
-                let outputs = self
-                    .onnx
-                    .run(tvec!(
-                        input_ids.into(),
-                        attention_mask.into(),
-                        token_type_ids.into()
-                    ))
-                    .map_err(|e| {
-                        tracing::error!("{e}");
-                        e
-                    })
-                    .ok()?;
-
-                let logits = outputs[0]
-                    .to_array_view::<f32>()
-                    .map_err(|e| {
-                        tracing::error!("{e}");
-                        e
-                    })
-                    .ok()?;
-
-                let logits = logits.slice(tract_ndarray::s![.., 0, ..]);
-
-                let norm = logits.to_owned().norm();
-
-                let data = logits.to_owned() / norm;
-                Some(PointStruct {
-                    id: Some(PointId::from(uuid::Uuid::new_v4().to_string())),
-                    vectors: Some(data.as_slice().unwrap().to_vec().into()),
-                    payload: hashmap! {
-                        "lang".into() => lang_str.to_ascii_lowercase().into(),
-                        "file".into() => relative_path.to_string_lossy().as_ref().into(),
-                        "snippet".into() => chunk.into(),
-                    },
-                    ..Default::default()
-                })
-            })
-            .collect::<Vec<_>>();
-
-        if datapoints.len() > 0 {
-            debug!("updating docs with {} points", datapoints.len());
-            Handle::current().block_on(async {
-                let status = self.qdrant.upsert_points("documents", datapoints).await;
-                if status.is_ok() {
-                    debug!("successful upsert");
-                }
-            });
-        }
 
         trace!("writing document");
-
         #[cfg(feature = "debug")]
         let buf_size = buffer.len();
         writer.add_document(doc!(
